14.09 Using separated lib easyroute-core
