+++ README.md
@@ -36,6 +36,10 @@ Why you should try it?
 
 Please see [DOCUMENTATION](https://easyroute.lyoha.info). Router is still developing and new features are coming out frequently.
 
+### Under the hood
+This is Svelte bindings for [Easyroute](https://github.com/lyohaplotinka/easyroute) framework-agnostic
+library.
+
 #### Contact me:
 * [My website: https://lyoha.info/](https://lyoha.info/) (russian language!)
 * [Telegram: https://t.me/alexeysolovjov](https://t.me/alexeysolovjov)
