+++ demo-app/src/App.svelte
@@ -1,5 +1,5 @@
 <script>
-	import { RouterOutlet } from '../../lib/dist'
+	import { RouterOutlet } from '../../lib'
 	export let router = null
 </script>
