+++ demo-app/src/Pages/Index.svelte
@@ -1,5 +1,5 @@
 <script>
-    import RouterOutlet from '../../../lib/svelte-components/RouterOutlet.svelte'
+    import RouterOutlet from '../../../lib'
 </script>
 
 <article>
