+++ lib/__tests__/Parsing.spec.js
@@ -1,4 +1,4 @@
-import Router from '../dist'
+import Router from 'easyroute-core'
 
 const config = {
   mode: 'history',
