+++ lib/index.js
@@ -1,0 +1,5 @@
+import Router from 'easyroute-core'
+
+export default Router
+export { default as RouterOutlet } from './RouterOutlet.svelte'
+export { default as RouterLink } from './RouterLink.svelte'
