--- lib/src/Utils/IdGenerator.ts
@@ -1,3 +1,0 @@
-export default function generateId() {
-  return Math.random().toString(36).substr(2, 9)
-}
