--- lib/src/index.ts
@@ -1,3 +1,0 @@
-import Router from './Router/Router'
-
-export default Router
